+++ Runtime/CinemachineInputSensitivity.cs
@@ -1,9 +1,4 @@
-#if CINEMACHINE_3
 using Unity.Cinemachine;
-#else
-using Cinemachine;
-#endif
-using System.Collections;
 using System.Collections.Generic;
 using UnityEngine;
 
@@ -11,60 +6,41 @@ namespace CharacterMovement
 {
     public class CinemachineInputSensitivity : MonoBehaviour
     {
-        private enum CameraInputMode
-        {
-            None,
-            FreeLook,
-            POV
-        }
-
         [field: SerializeField] private float Multiplier { get; set; } = 1f;
-        [field: SerializeField] private float TargetFPS { get; set; } = 60f;
 
-        private Vector2 _sensitivity;
-        private CinemachineFreeLook _freeLook;
-        private CinemachinePOV _pov;
-        private CameraInputMode _mode;
+        private List<InputAxisControllerBase<CinemachineInputAxisController.Reader>.Controller> _controllers;
+        private float[] _originalGains;
 
         private void Start()
         {
-            if (TryGetComponent(out _freeLook))
+            if(TryGetComponent(out CinemachineInputAxisController controllerComponent))
             {
-                _sensitivity.x = _freeLook.m_XAxis.m_MaxSpeed;
-                _sensitivity.y = _freeLook.m_YAxis.m_MaxSpeed;
-                _mode = CameraInputMode.FreeLook;
-            }
-            else if (TryGetComponent(out CinemachineVirtualCamera vcam))
-            {
-                _pov = vcam.GetCinemachineComponent<CinemachinePOV>();
-                if (_pov != null)
+                _controllers = controllerComponent.Controllers;
+                _originalGains = new float[_controllers.Count];
+                for (int i = 0; i < _controllers.Count; i++)
                 {
-                    _sensitivity.x = _pov.m_HorizontalAxis.m_MaxSpeed;
-                    _sensitivity.y = _pov.m_VerticalAxis.m_MaxSpeed;
-                    _mode = CameraInputMode.POV;
+                    _originalGains[i] = _controllers[i].Input.Gain;
                 }
+
+                SetMultiplierAll(Multiplier);
             }
         }
 
-        private void Update()
+        public void SetMultiplierAll(float multiplier)
         {
-            float deltaTime = Mathf.Clamp(Time.smoothDeltaTime, 0.001f, 1f);
-            float currentFPS = 1f / deltaTime;
-            Vector2 adjustedSensitivity = _sensitivity * currentFPS / TargetFPS * Multiplier;
-
-            switch (_mode)
+            for (int i = 0; i < _controllers.Count; i++)
             {
-                case CameraInputMode.None:
-                    break;
-                case CameraInputMode.FreeLook:
-                    _freeLook.m_XAxis.m_MaxSpeed = adjustedSensitivity.x;
-                    _freeLook.m_YAxis.m_MaxSpeed = adjustedSensitivity.y;
-                    break;
-                case CameraInputMode.POV:
-                    _pov.m_HorizontalAxis.m_MaxSpeed = adjustedSensitivity.x;
-                    _pov.m_VerticalAxis.m_MaxSpeed = adjustedSensitivity.y;
-                    break;
+                SetMultiplier(i, multiplier);
             }
+        }
+
+        public void SetMultiplier(int controllerIndex, float multiplier)
+        {
+            if(_controllers == null || _controllers.Count <= controllerIndex || _originalGains == null || _originalGains.Length <= controllerIndex)
+            {
+                Debug.LogError($"CinemachineInputAxisController axis missing at Index: {controllerIndex}");
+            }
+            _controllers[controllerIndex].Input.Gain = _originalGains[controllerIndex] * multiplier;
         }
     }
 }